FirstDjangoProject/requirements.txt
asgiref==3.7.2
astroid==3.0.1
black==23.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
Django==4.2.5
django-reversion==5.0.4
django-simple-history==3.4.0
djangorestframework==3.14.0
graphviz==0.20.1
idna==3.6
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
py2puml==0.8.1
pylint==3.0.2
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.2
urllib3==2.1.0
